Update virkill.ps1
VirKill no longer needs to have a runme.bat to run. I have condensed everything in the bat file to the powershell script. An external script can be made to launch the powershell script. As a note, this can not be packaged into an EXE with PS2EXE. It throws errors about the path, but I am actively looking into fixing this problem.